% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{CoG}
\alias{CoG}
\title{Covariate Set in the G->X path}
\format{A set with 2000 rows and 5 variables:
\describe{
  \item{C1 - C3}{Three continuous covariates}
  \item{C4, C5}{Two binary covariates}
}}
\usage{
CoG
}
\description{
A simulated dataset containing one of the optional components to run est_lucid,
plot_lucid, and tune_lucid. The variables are as follows:
}
\keyword{datasets}
