% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{CoY}
\alias{CoY}
\title{Covariate Set in the X->Y path}
\format{A set with 2000 rows and 5 variables:
\describe{
  \item{YC1 - YC3}{Three continuous covariates}
  \item{YC4, YC5}{Two binary covariates}
}}
\usage{
CoY
}
\description{
A simulated dataset containing one of the optional components to run est_lucid,
plot_lucid, and tune_lucid. The variables are as follows:
}
\keyword{datasets}
