% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/est_cluster.R
\name{est_cluster}
\alias{est_cluster}
\title{Estimating latent clusters with multi-omics data}
\usage{
est_cluster(G = NULL, Z = NULL, Y, family = "binary", useY = TRUE,
  init_b = NULL, init_m = NULL, init_s = NULL, init_g = NULL,
  init_pcluster = NULL, K = 2, MAX_ITR = 100, MAX_TOT_ITR = 10000,
  reltol = 1e-08, tol_b = 1e-04, tol_m = 1e-04, tol_s = 1e-04,
  tol_g = 1e-04, tol_p = 1e-04, Pred = FALSE, Select_G = FALSE,
  Select_Z = FALSE, Rho_G = -9, Rho_Z_InvCov = 0, Rho_Z_CovMu = 0)
}
\arguments{
\item{G}{Genetic effects, a matrix}

\item{Z}{Biomarker data, a matrix}

\item{Y}{Disease outcome, a vector}

\item{family}{"binary" or "normal" for Y}

\item{useY}{Using Y or not, default is TRUE}

\item{K}{Pre-specified # of latent clusters, default is 2}

\item{MAX_ITR}{Maximum number of iterations, integer, default is 100}

\item{MAX_TOT_ITR}{Maximum number of total iterations, integer, default is 10000}

\item{reltol}{Convergence cut-off using a relative tolerance, default is 1e-8}

\item{Pred}{Flag to compute predicted disease probability with fitted model, boolean, default is FALSE}

\item{Select_G}{Flag to do model selection on genetic data, default is FALSE}

\item{Select_Z}{Flag to do model selection on biomarker data, default is FALSE}

\item{Rho_G}{Penalty for selection on genetic data, numeric, default is -9 using a sequence of penalties}

\item{Rho_Z_InvCov}{Penalty for the inverse of covariance of biomarkers, numeric, default is 0}

\item{Rho_Z_CovMu}{Penalty for the product of covariance and mean of biomarkers, numeric, default is 0}
}
\value{
\code{est_cluster} returns an object of list containing parameters estimates, predicted probability of latent clusters, and other features:
\item{beta}{Estimates of genetic effects, matrix}
\item{mu}{Estimates of cluster-specific biomarker means, matrix}
\item{sigma}{Estimates of cluster-specific biomarker covariance matrix, list}
\item{gamma}{Estimates of cluster-specific disease risk, vector}
\item{pcluster}{Probability of cluster, when G is null}
\item{pred}{Predicted probability of belonging to each latent cluster}
}
\description{
\code{est_cluster} estimates an integrated cluster assignment of genetic effects using complete biomarker data with/without disease outcomes. Options to produce sparse solutions for cluster-specific parameter estimates under a circumstance of analyzing high-dimensional data are also provided.
}
\examples{
# For a testing dataset with 10 genetic features (5 causal) and 4 biomarkers (2 causal)
est_cluster(G=G1,Z=Z1,Y=Y1,K=2,
            init_b = NULL, init_m = NULL, init_s = NULL, init_g = NULL,
            family="binary",Pred=TRUE,Select_G=TRUE,Select_Z=TRUE,
            Rho_G=0.02,Rho_Z_InvCov=0.1,Rho_Z_CovMu=93,
            tol_m = 1e-8,tol_b=1e-8,tol_s=1e-8,tol_g=1e-8,MAX_ITR = 800,MAX_TOT_ITR=800)
}
\references{
Peng, C., Yang, Z., Conti, D.V.
}
\author{
Cheng Peng, Zhao Yang, David V. Conti
}
\keyword{cluster}
\keyword{latent}
