% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_lucid.R
\name{plot_lucid}
\alias{plot_lucid}
\title{Plot Sankey diagram for integrative clustering}
\usage{
plot_lucid(x, colorScale = default)
}
\arguments{
\item{x}{An \code{IntClust} class object}

\item{colorScale}{D3 color scheme for the Sankey diagram}
}
\description{
\code{plot_lucid} generates a Sankey diagram for the results of integrative clustering based on an \code{IntClust} object.
}
\references{
Peng, C., Yang, Z., Conti, D.V.
}
\author{
Cheng Peng, Zhao Yang, David V. Conti
}
