% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sem_cluster.R
\name{sem_cluster}
\alias{sem_cluster}
\title{SEM for latent cluster estimation}
\usage{
sem_cluster(G = NULL, Z = NULL, Y, family = "binary", useY = TRUE,
  init_b = NULL, init_m = NULL, init_s = NULL, init_g = NULL,
  init_pcluster = NULL, K = 2, MAX_ITR = 100, MAX_TOT_ITR = 10000,
  reltol = 1e-08, tol_b = 1e-04, tol_m = 1e-04, tol_s = 1e-04,
  tol_g = 1e-04, tol_p = 1e-04, tol_sem = 1e-04, Pred = FALSE,
  Get_SE = TRUE, Ad_Hoc_SE = FALSE)
}
\arguments{
\item{G}{Genetic effects, a matrix}

\item{Z}{Biomarker data, a matrix}

\item{Y}{Disease outcome, a vector}

\item{family}{"binary" or "normal" for Y}

\item{useY}{Using Y or not, default is TRUE}

\item{K}{Pre-specified # of latent clusters, default is 2}

\item{MAX_ITR}{Maximum number of iterations, integer, default is 100}

\item{MAX_TOT_ITR}{Maximum number of total iterations, integer, default is 10000}

\item{reltol}{Convergence cut-off using a relative tolerance, default is 1e-8}

\item{Pred}{Flag to compute predicted disease probability with fitted model, boolean, default is FALSE}

\item{Get_SE}{Flag to perform SEM to get SEs of parameter estimates, default is TRUE}

\item{Ad_Hoc_SE}{Flag to fit ad hoc regression models to get SEs of parameter estimates, default is FALSE}
}
\value{
\code{sem_cluster} returns an object of list containing parameters estimates, their corresponding standard errors, and other features:
\item{beta}{Estimates of genetic effects, matrix}
\item{se_beta}{SEM standard errors of Beta}
\item{se_ah_beta}{Ad hoc standard errors of Beta}
\item{mu}{Estimates of cluster-specific biomarker means, matrix}
\item{se_mu}{SEM standard errors of Mu}
\item{se_ah_mu}{Ad hoc standard errors of Mu}
\item{sigma}{Estimates of cluster-specific biomarker covariance matrix, list}
\item{gamma}{Estimates of cluster-specific disease risk, vector}
\item{se_gamma}{SEM standard errors of Gamma}
\item{se_ah_gamma}{Ad hoc standard errors of Gamma}
\item{pcluster}{Probability of cluster, when G is null}
\item{pred}{Predicted probability of belonging to each latent cluster}
}
\description{
\code{sem_cluster} provides standard errors (SE) of parameter estimates when performing latent cluster analysis with multi-omics data. SEs are obtained through supplemented EM-algorithm (SEM).
}
\examples{
sem_cluster(G=G2,Z=Z2,Y=Y2,useY=TRUE,K=2,
            init_b = NULL, init_m = NULL, init_s = NULL, init_g = NULL,
            Pred=TRUE,family="normal",Get_SE=TRUE,Ad_Hoc_SE=TRUE,
            tol_m = 1e-6,tol_b=1e-6,tol_s=1e-6,tol_g=1e-6,tol_sem=1e-3,
            MAX_ITR=3000,MAX_TOT_ITR=3000)
}
\references{
Meng, X., & Rubin, D. B. (1991). Using EM to Obtain Asymptotic Matrices : The SEM Algorithm.
Journal of the American Statistical Association, 86(416), 899-909. http://doi.org/10.2307/2290503
}
\author{
Cheng Peng, Zhao Yang, David V. Conti
}
\keyword{SEM,}
\keyword{cluster}
\keyword{latent}
