% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tune_cluster.R
\name{tune_cluster}
\alias{tune_cluster}
\title{Parallel Grid Search for Tuning Parameters in Latent Cluster Analysis}
\usage{
tune_cluster(G = NULL, Z = NULL, Y, K, Family, USEY = TRUE,
  start_b = NULL, start_m = NULL, start_s = NULL, start_g = NULL,
  LRho_g, URho_g, NoRho_g, LRho_z_invcov, URho_z_invcov, NoRho_z_invcov,
  LRho_z_covmu, URho_z_covmu, NoRho_z_covmu)
}
\arguments{
\item{G}{Genetic effects, a matrix}

\item{Z}{Biomarker data, a matrix}

\item{Y}{Disease outcome, a vector}

\item{K}{Pre-specified # of latent clusters, default is 2}

\item{Family}{"binary" or "normal" for Y}

\item{USEY}{Using Y or not, default is TRUE}

\item{LRho_g}{Lower limit of the penalty for selection on genetic data}

\item{URho_g}{Upper limit of the penalty for selection on genetic data}

\item{NoRho_g}{Number of Rho_g for grid-search}

\item{LRho_z_invcov}{Lower limit of the penalty for the inverse of covariance of biomarkers}

\item{URho_z_invcov}{Upper limit of the penalty for the inverse of covariance of biomarkers}

\item{NoRho_z_invcov}{Number of Rho_z_invcov for grid-search}

\item{LRho_z_covmu}{Lower limit of the penalty for the product of covariance and mean of biomarkers}

\item{URho_z_covmu}{Upper limit of the penalty for the product of covariance and mean of biomarkers}

\item{NoRho_z_covmu}{Number of Rho_z_covmu for grid-search}
}
\value{
\code{tune_cluster} returns an object of list containing Modelfits, Results, and Optimal:
\item{Modelfits}{Latent cluster model fits for a combination of given tuning parameters}
\item{Results}{Summary results of grid-search}
\item{Optimal}{Fatures of the optimal model with minimun BIC in the grid-search results}
}
\description{
\code{tune_cluster} fits regularized latent cluster models with various combinations of three tuning parameters based on joint inference across data types to perform a grid-search helping determine an optimal choice of three tuning parameters with minimum model BIC.
}
\examples{
# For a testing dataset with 10 genetic features (5 causal) and 4 biomarkers (2 causal)
# Assign the number of cores
no_cores <- detectCores() - 1
# Initiate cluster
cl <- makeCluster(no_cores)
#Start parallel computing
registerDoParallel(cl)
GridSearch <- tune_cluster(G=G1, Z=Z1, Y=Y1, K=2, Family="binary", USEY = TURE,
                           LRho_g = 0.01, URho_g = 0.03, NoRho_g = 3,
                           LRho_z_invcov = 0.1, URho_z_invcov = 0.2, NoRho_z_invcov = 2,
                           LRho_z_covmu = 91, URho_z_covmu = 95, NoRho_z_covmu = 5)
stopCluster(cl)
GridSearch$Results
GridSearch$Optimal
}
\references{
Peng, C., Yang, Z., Conti, D.V.
}
\author{
Cheng Peng, Zhao Yang, David V. Conti
}
\keyword{Grid-search}
\keyword{Parameter,}
\keyword{Tunning}
