% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{Z1_Incomp}
\alias{Z1_Incomp}
\title{Incomplete Biomarker Set 1}
\format{A set with 2000 rows and 4 variables, 500 rows are NAs:
\describe{
  \item{CZ1, CZ2}{Causal biomarkers}
  \item{NZ1, NZ2}{Null biomarkers}
}}
\usage{
Z1_Incomp
}
\description{
A simulated dataset containing one of the components to run est_lucid incomplete
option, plot_lucid, and tune_lucid. The variables are as follows:
}
\keyword{datasets}
