% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/def_tol.R, R/def_tune.R
\name{def_tol}
\alias{def_tol}
\alias{def_tol}
\title{Define maximum number of iteration and convergence}
\usage{
def_tol(Rho_G = -9, Rho_Z_InvCov = 0, Rho_Z_CovMu = 0, Select_G = FALSE,
  Select_Z = FALSE)

def_tol(Rho_G = -9, Rho_Z_InvCov = 0, Rho_Z_CovMu = 0, Select_G = FALSE,
  Select_Z = FALSE)
}
\arguments{
\item{Rho_G}{Penalty for selection on genetic data, numeric, default is -9 using a sequence of penalties}

\item{Rho_Z_InvCov}{Penalty for the inverse of covariance of biomarkers, numeric, default is 0}

\item{Rho_Z_CovMu}{Penalty for the product of covariance and mean of biomarkers, numeric, default is 0}

\item{Select_G}{Flag to do model selection on genetic data, default is FALSE}

\item{Select_Z}{Flag to do model selection on biomarker data, default is FALSE}

\item{MAX_ITR}{Maximum number of iterations, integer, default is 100}

\item{MAX_TOT_ITR}{Maximum number of total iterations, integer, default is 10000}

\item{reltol}{Convergence cut-off using a relative tolerance, default is 1e-8}

\item{tol_b}{Convergence criteria of \eqn{\beta}, genetic effects parameter, default is 1e-4}

\item{tol_m}{Convergence criteria of \eqn{\mu}, biomarker mean effects parameters, default is 1e-4}

\item{tol_s}{Convergence criteria of \eqn{\Sigma}, biomarker covariance matrix, default is 1e-4}

\item{tol_g}{Convergence criteria of \eqn{\gamma}, outcome effects parameter, default is 1e-4}

\item{tol_p}{Convergence criteria of the probability of latent clusters, default is 1e-4}
}
\value{
A list of tolerance settings will be returned for integrative clustering.

A list of tuning parameters and settings will be returned for integrative clustering.
}
\description{
Defines tolerance settings in \code{\link{est_cluster}}, \code{\link{sem_cluster}} , & \code{\link{tune_cluster}} fitting.

Defines selection options and tuning parameters in \code{\link{est_cluster}}, \code{\link{sem_cluster}} fitting.
}
\author{
Cheng Peng, Zhao Yang, David V. Conti

Cheng Peng, Zhao Yang, David V. Conti
}
