% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_lucid.R
\name{plot_lucid}
\alias{plot_lucid}
\title{Plot Sankey diagram for integrative clustering}
\usage{
plot_lucid(x, ...)
}
\arguments{
\item{x}{An \code{IntClust} class object}

\item{...}{Additional graphics parameters}
}
\description{
\code{plot_lucid} generates a Sankey diagram for the results of integrative clustering based on an \code{IntClust} object.
}
\references{
Peng, C., Yang, Z., Conti, D.V.
}
\author{
Cheng Peng, Zhao Yang, David V. Conti
}
