% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary_lucid.R
\name{summary_lucid}
\alias{summary_lucid}
\title{Summarize results for integrative clustering}
\usage{
summary_lucid(x, ...)
}
\arguments{
\item{x}{An \code{IntClust} class object}

\item{...}{Additional parameters}
}
\value{
\code{summary_lucid} returns a list containing important outputs from an \code{IntClust} object.
\item{Beta}{Estimates of genetic effects, matrix}
\item{Mu}{Estimates of cluster-specific biomarker means, matrix}
\item{Gamma}{Estimates of cluster-specific disease risk, vector}
\item{select_G}{A logical vector indicates non-zero genetic features}
\item{select_Z}{A logical vector indicates non-zero bio-features}
\item{No0G}{A total # of non-zero genetic features}
\item{No0Z}{A total # of non-zero bio-features}
\item{BIC}{Model BIC}
}
\description{
\code{summary_lucid} generates a summary for the results of integrative clustering based on an \code{IntClust} object.
}
\references{
Peng, C., Yang, Z., Conti, D.V.
}
\author{
Cheng Peng, Zhao Yang, David V. Conti
}
